Adds InlineLoader output
