+++ webpack.config.js
@@ -1,3 +1,3 @@
-const { production, debug } = require("./webpackConfigs.js");
+const { production, debug, debugInlineLoader, prodInlineLoader } = require('./webpackConfigs.js')
 
-module.exports = [production, debug];
+module.exports = [production, debug, debugInlineLoader, prodInlineLoader]
